conn: raw []byte, not msg
This commit actually removes the previously introduced
chan net.NetMessage, in favor of raw []byte. It plays
nicer with crypto/spipe, and it makes more sense in the
context of a "single connection", i.e. I already know the
peer I'm talking to, from the connection. The NetMessage
additional Peer is useful swarm and up.