feat(docker) build a daemon
The default Docker build creates a daemon.

TODO allow bootstrap node to be overridden at runtime (perhaps with an
environment variable), so we can use custom topologies in test
environments. NB: We still want the default behavior to connect to Mars.

TODO create a docker image that mounts and uses shared volume to make
that mount accessible on the host.