conn: larger max message.
With the new msgio, we no longer need to bound message sizes for
the sake of buffers. MaxMessageSize should remain though in order
to prevent streaming of huge things. In general, messages shouldn't
be above one MB. This max is now set to 4MB.

If a block happens to be larger, we should split it up.