fix fuse mount error in linux
There has been a regression such that ./t0030-mount.sh fails on

  'ipfs mount' fails when there is no mount dir

The issue was a change in how fuse errors are reported to the client
process. We have introduced an optimistic categorization that hides
the obscure fusermount error and replaces it with something a bit
more helpful.

License: MIT
Signed-off-by: Juan Batiz-Benet <juan@benet.ai>