Moved Tests up in Contribute.md
This should be more prominent. We also need a much better description of how to run tests - rigt now, we are dependant on Travis for the majority of testing, I believe. If there was a clear guide here on how to run the appropriate tests for each kind of PR, that would be fantastic.

We also really ought to have a section asking people to add all of the tests that they can think of.

License: MIT
Signed-off-by: Richard Littauer <richard.littauer@gmail.com>