sharness: don't run go-ipfs instance with default config
It conflicts with already running instance on the Dev's machine so
running this test requires disabling the normal ipfs.

The launch of ipfs here doesn't test much as we compare the config,
after initing from it and we also launch from default config in many
other cases (with few variables changed to not conflict with already
running standalone ipfs instance).

License: MIT
Signed-off-by: Jakub Sztandera <kubuxu@protonmail.ch>