make sure the repo size is greater than the size of the symlink
Before, we'd check to make sure the repo, when checked through a symlink, is at
least as large as the repo *before* we checked it through the symlink. However,
this assumes that the repo can't shrink.

Really, this test exists to ensure we measure the repo size itself instead of
the size of the symlink; this commit changes the test to reflect this.

This test fails when 54d7e03303f70d88bfaa8a8aee636dd33b2dd491 is reverted.

fixes #4408

License: MIT
Signed-off-by: Steven Allen <steven@stebalien.com>