re-format log output as ndjson
Apparently, js-ipfs-api expects this. Really, I'm not sure why js-ipfs-api even
cares about this endpoint but that's a different issue.

The real fix is to change back to ndjson in go-log but that would force us to
update every package (not good when we're trying to push out a release like
this).

License: MIT
Signed-off-by: Steven Allen <steven@stebalien.com>