fix: show interactive output from install.sh
When trying out the latest rc, i used the `./install.sh`
It hung, with no output. I removed the 2> /dev/null and
it turned out mv was waiting for user input to confirm
the change of file permssions from 555 to 755.

This PR removes piping the output from mv to /dev/null
as it seems like the safest fix.

An alternative would be to just add -f but i've err'd
on the side of caution.

If also tweaked the second condition basedo on the
recommendations of shellcheck
see: https://github.com/koalaman/shellcheck/wiki/SC2166

License: MIT
Signed-off-by: Oli Evans <oli@tableflip.io>